+++ packages/core-events/src/module/configureEventsModule.ts
@@ -1,13 +1,11 @@
-import type { mongodb } from '@unchainedshop/mongodb';
-
 import { generateDbFilterById, generateDbMutations, buildSortOptions } from '@unchainedshop/mongodb';
 import { getRegisteredEvents } from '@unchainedshop/events';
 import { SortDirection, SortOption } from '@unchainedshop/types/api.js';
-import { ModuleCreateMutation, ModuleInput, ModuleMutations } from '@unchainedshop/types/core.js';
+import { ModuleInput, ModuleMutations } from '@unchainedshop/types/core.js';
 import { EventsCollection, Event } from '../db/EventsCollection.js';
 import { EventsSchema } from '../db/EventsSchema.js';
 import { configureEventHistoryAdapter } from './configureEventHistoryAdapter.js';
-import { EventReport } from '@unchainedshop/types/events.js';
+import { EventReport, EventsModule } from '@unchainedshop/types/events.js';
 
 export type EventQuery = {
   types?: Array<string>;
@@ -23,27 +21,6 @@ export const buildFindSelector = ({ types, queryString, created }: EventQuery) =
   if (created) selector.created = { $gte: created };
   return selector;
 };
-
-export interface EventsModule extends ModuleCreateMutation<Event> {
-  findEvent: (
-    params: mongodb.Filter<Event> & { eventId: string },
-    options?: mongodb.FindOptions,
-  ) => Promise<Event>;
-
-  findEvents: (
-    params: EventQuery & {
-      limit?: number;
-      offset?: number;
-      sort?: Array<SortOption>;
-    },
-    options?: mongodb.FindOptions,
-  ) => Promise<Array<Event>>;
-
-  type: (event: Event) => string;
-
-  count: (query: EventQuery) => Promise<number>;
-  getReport: (params?: { from?: Date; to?: Date; types?: string[] }) => Promise<EventReport[]>;
-}
 
 export const configureEventsModule = async ({
   db,
@@ -84,22 +61,22 @@ export const configureEventsModule = async ({
       return count;
     },
 
-    getReport: async ({ from, to, types } = { from: null, to: null, types: null }) => {
+    getReport: async ({ dateRange, types } = { dateRange: {}, types: null }) => {
       const pipeline = [];
       const matchConditions = [];
       // build date filter based on provided values it can be a range if both to and from is supplied
       // a upper or lowe limit if either from or to is provided
       // or all if none is provided
-      if (from || to) {
+      if (dateRange?.start || dateRange?.end) {
         const dateConditions = [];
-        if (from) {
-          const fromDate = new Date(from);
+        if (dateRange?.start) {
+          const fromDate = new Date(dateRange?.start);
           dateConditions.push({
             $or: [{ created: { $gte: fromDate } }, { updated: { $gte: fromDate } }],
           });
         }
-        if (to) {
-          const toDate = new Date(to);
+        if (dateRange?.end) {
+          const toDate = new Date(dateRange?.end);
           dateConditions.push({
             $or: [{ created: { $lte: toDate } }, { updated: { $lte: toDate } }],
           });